Don't free a cookie we passed to netpacket_op.
The netpacket_op code is peculiar in that it can fail in multiple ways,
with different results: If netpacket_op fails to add the request to its
internal queue, it will return with the request not queued (obviously),
but if netpacket_op fails at a later point, the request remains on the
queue and may be handled (and a callback made) later.

In order to ensure that we do not use a memory allocation after freeing
it, don't free our read request cookie if netpacket_op fails.  This may
result in the memory allocation being leaked; that is the lesser of the
two potential bugs.

(It turns out that this is harmless either way, since tarsnap will die
promptly if storage_read_file_callback ever fails.  But better to opt
for the less harmful bug in case that ever changes.)

At some point netpacket_op should probably be fixed to avoid returning
failure with a request queued; at which point this commit can be
reverted (since it will then be safe to free our cookie upon seeing a
failure from netpacket_op).