Don't add corrupt data to the block cache.
Check that the status code is still 0 (success) before adding a block
to the cache.  While we're here, set C->buf to NULL upon freeing it,
as an extra seatbelt against use-after-free bugs.

This was harmless since a data-is-corrupt read failure causes tarsnap
to abort; so there is no potential for the corrupt data to ever be
read out of the cache.