Make `tarsnap --version` exit with non-failure.
Since `tarsnap --version` is a deliberate request for the version
number, it should yield an exit status of zero (success).

This logic does not apply to the "help" information which is printed
in response to `tarsnap --bogus-option`, since we need to exit with
a failure status if the user may have typoed an option name.