Add warning if user tries to archive --foo
A frequent mistake made by tarsnap users is to attempt to place options
at the end of the command line, e.g.

This does not work, since UNIX utilities process the command line from
left to right and stop looking for options as soon as they first
non-option argument is found.

While someone might want to deliberately archive the "--exclude" file,
it is unlikely; so we proceed but print a warning first.