Drop unnecessary support from archive read path.
I originally kept archive_read_support_{format, compression}_all in the
tarsnap archive read path in order to provide compatibility in case I
changed the internal format of tarsnap archives.  At this point that
seems unlikely; and supporting "unexpected" types of archives increases
the attack surface by which someone with the archive-creation keys could
attack a system which is extracting or listing the contents of archives.

In short: Adjust convenience/security tradeoff further in the direction
of paranoia.