Don't crash if initial DNS lookup fails.
Tarsnap keeps the result of the last successful DNS lookup cached, so
that it can be reused in the event that a future lookup fails.  So far
so good.  Unfortunately in tarsnap 1.0.34, as part of code refactoring
related to gaining support for IPv6, this code was changed to copying
the address rather than merely passing a pointer to the cached address.

Since, if the initial DNS lookup fails, the cached address will be NULL,
this copying did not go very well.  Instead, only copy the address if
non-NULL; return a (uncopied) NULL otherwise.