MFlibcperciva r50:
------------------------------------------------------------------------
r50 | cperciva | 2013-11-12 00:06:12 -0800 (Tue, 12 Nov 2013) | 3 lines

If connect() returns -1 and errno is not EINPROGRESS or EINTR, the request
failed.  "(errno != EINPROGRESS) && (errno == EINTR)" was a weird typo.

------------------------------------------------------------------------

The effect of this bug on tarsnap was to cause tarsnap to error out if
the first address (IPv4 or IPv6) it tried could not connect; most often
this happened with IPv6-only networks.