MFlibcperciva r49
------------------------------------------------------------------------
r49 | cperciva | 2013-11-11 23:16:21 -0800 (Mon, 11 Nov 2013) | 6 lines

If clock_gettime(CLOCK_MONOTONIC) fails with ENOSYS or EINVAL, fall back
to using gettimeofday().  There are platforms on which CLOCK_MONOTONIC is
defined in spite of not actually being supported by the kernel; POSIX
specifies EINVAL for this, but there are some reports of ENOSYS being used
instead, so accept either to trigger the fallback.
------------------------------------------------------------------------