Distrust 32-bit hw.usermem sysctls.
Some platforms return a 32-bit value for hw.usermem even though they
support more than 4GB of RAM.  This is bad enough already; but they can
truncate the amount of memory modulo 4GB, yielding zero or absurdly
small values.  As a result, it's better to simply ignore 32-bit results
from hw.usermem.

If hw.usermem returns an 8-byte value, we continue to parse that as a
64-bit integer and use that value.