Correct number of keys; avoid UB.
Tarsnap keys are represented as bits in a 32-bit word; so there
is no key #32, and we should complain if we are asked for it.

Shifting by more than the length of an integer is undefined
behaviour, so (now that we've guaranteed that we'll only have
indices in the range 0 .. 31) make explicit that we're shifting
a 32-bit "1" to construct the appropriate bitmask.