Hook AESNI support into the build.
This builds AESNI-enabled AES routines and the cpu-detection code which
they require.  Note that these routines are presently unused.

Due to autotools stupidity -- you can't build separate files in a
library with different compiler flags -- it's necessary to build the
AESNI routines into a separate library which gets included everywhere.
(We can't pass the AESNI-enabling compiler flags when compiling all
the other files, since that would permit the compiler to use AESNI
instructions in the "we've detected that we're running on a CPU which
does not support AESNI" code paths.)  This is ugly right now, but will
become quite tiresome when there are more CPU features...