Add --retry-forever option.
This option, if set, causes tarsnap to continue trying to reconnect to
the tarsnap server forever, instead of giving up after 5-10 minutes.
This option may be useful for people with excessively flaky networks
or on mobile devices which regularly lose their internet connections
for extended periods of time; it will however not be turned on by
default since the combination of a network (mis)configuration which
breaks tarsnap and a system administrator who does not verify that
backups have completed could result in a tarsnap process "lingering"
for an extended period of time without overtly failing.

This option has not been added to the tarsnap man page yet.  That
should happen before the next release.