Remove unnecessary check against (size_t)(-1)
Between 2007-10-09 and 2007-10-17 the multitape_metadata_put function
returned the size of the stored metadata as a size_t, or returned
(size_t)(-1) on error.  In order to disambiguate these cases, I added
a check for buflen == (size_t)(-1) and errored out in that case.  (In
practice it was impossible for buflen to be that large anyway.)

When I changed multitape_metadata_put to follow coding style and
return an int for status, I left behind this code which was no longer
necessary; and it took until now before anyone drew this code to my
attention.

Noticed by:	gperciva