Use cached file flags on linux for nodump check.
This avoids opening and re-reading the file flags on filesystem
objects.  This may make things slightly faster; but far more
importantly, it avoids opening things which are not directories
or regular files.

This was causing problems when tarsnap recursed into /dev and
examined /dev/watchdog: Because opening that device starts the
watchdog timer automatically (!) it would result in the system
rebooting a short time later.  With this change, filesystem
objects which are not directories or regular files are never
opened, and the watchdog is not triggered.

(The code in write.c could have been adjusted by simply adding a
check for the filesystem object type and skipping the nodump check
for anything other than directories and regular files -- this is
what the code in libarchive which reads the file flags does -- but
since the file flags are available from libarchive it's simpler to
just use those.)

Reported by:	Benjamin Gilbert
Bug Bounty:	$100