Remove chunkification cache when --fsck is run
Tarsnap's chunkification cache exists as an optimization; if files have
not changed since the last time they were archived, the cache may make
it possible to save disk I/O and CPU time by not reading the files and
splitting them into chunks again.

That said, it is *only* an optimization; it can be safely deleted at
any time.

Sometimes when `tarsnap --fsck` is run, it is because the client has
fallen out of sync with the server thanks to filesystem corruption.  In
such an event, it's possible that the chunkification cache has also
been damaged; so out of an abundance of caution we will remove the
potentially suspect cache -- trading a slight increase in CPU and disk
I/O usage for eliminating a possible route whereby hardware failure
could result in data loss.

X-Github:	Closes #16