Fix one-byte buffer overflow
If the first path encountered by tarsnap which is >= 1024/2048/4096/etc
bytes in length is exactly 1024/2048/4096/etc bytes in length, then the
path buffer will be overflowed by a single NUL byte.  This can crash
tarsnap (by overflowing into an unallocated page); it might also be
possible for this buffer overflow to be exploited to cause something
worse than a crash, but it seems like this would be challenging.

Reported by:	Ariel Ben Yehuda
Bug Bounty:	$100