Refactor key*gen: share all networking code
This moves material from the main() of key*gen.c into keygen_network.c, except
that it replaces "C." with "C->" (since we pass C as a pointer).

The err1, err2 labels do not follow Tarsnap STYLE; they should have 1
subtracted from them.  This will be adjusted in the next commit.  (They were
not adjusted in this commit so that each commit would be a minimal step, and
thus easier to check.)