key*gen: stricter check for quick callback_register_response()
If a misbehaving server replied with NETPACKET_REGISTER_RESPONSE with the
status being "Success" or "Account balance is not positive", before issuing a
NETPACKET_REGISTER_CHALLENGE, the challenge response would be skipped and
crypto_hash_data_key_2() will be called with an uninitialized key.  This is
technically undefined behaviour (doing arithmetic on uninitialized memory).

However, the end result is merely that the user would receive a
    tarsnap-keygen: Network protocol violation by server
message.  This could arise in two ways:
- after calling crypto_hash_data_key_2() with an uninitialized key (from
  C->register_key), the result is then subjected to a crypto_verify_bytes()
  check.  If it fails, the program exits with "protocol violation" message.
- if crypto_verify_bytes() somehow succeeds, then program flow returns to
  keygen_network_register().  However, the
    if (C->donechallenge == 0) && (C->status != 1))
  check reveals that something is wrong; C->donechallenge is only set if the
  challenge is passed.  The program then exits with the "protocol violation"
  message.

Either way, the impact of this bug is simply that we exit with a "protocol
violation" error a few microseconds later than we should have.

Reported by:	Anderson Lizardo
Bug bounty:	$20