Re-hash disk data and compare with cached hashes
This could provide some warning about a failing disk.

For debugging: to create a modified file without altering the mtime (i.e.
faking a disk error):

    cp -p good temp
    vi temp
    touch -r good temp
    cp -p temp good

Invoking bsdtar_errc() leaves some memory which was not properly free()'d.  As
far as I can tell, this is unavoidable without lots of changes to libarchive
code -- too many of their functions are 'void' and have no mechanism for
reporting errors.  :(   This is something that I'm interested in improving at
some point the future, but there's more important things to focus on right
now.