Build: if bash completion <= 2.0, set a default dir
Previously if the system did not have bash-completion and the user asked for
an automatic directory, the configure script would quit.  Now, it attempts to
continue by setting the directory to the value suitable for Ubuntu 12.04
(since I estimate this is a combination of "most popular" and "most likely to
have users who aren't comfortable specifying a directory in
--with-bash-completion-dir[=DIR]").

Note that FreeBSD ports upgraded to bash-completion 2.0 on 02 Nov 2012
(revision 306891), and debian wheezy 7.0 on 04 May 2013 uses bash-completion
2.0.  So it's only old servers that don't have 2.0.

The main reason I'm making this change is so that the deb packages can rely on
autodetection (and fallback on Ubuntu 12.04), instead of needing to hard-code
an old directory on modern distributions.