Use ttyfd to find the terminal for sigquit
The old code assumed that STDIN_FILENO was a descriptor for the
attached terminal, and attempted to call tc(get|set)attr on it.
This is not always the case; tarsnap may be at the end of a
pipeline, for example.

The (recently added) ttyfd code tries harder to find the attached
terminal.