Change server name from "betatest" to "v1-0-0"
Tarsnap ceased to be in beta testing a long time ago (it sort of
drifted into release without there being any official announcement)
but I never changed the server hostname.  For the sake of avoiding
confusion, I am changing it now; the old name will still work (and
points to the same addresses) but tarsnap will now look up the host
"v1-0-0-server.tarsnap.com".

If it becomes necessary for future versions of tarsnap to contact a
different set of addresses (e.g., due to a change in the protocol)
then new server hostnames may be used with "1-0-0" replaced by the
version of tarsnap where the change is made.  At present there is
no reason to think this will ever become necessary.