Work around LLVM bug #27190
The C99 and C11 standards state (7.13.2.1) that
"the values of objects
    1. of automatic storage duration
    2. that are local to the function containing the invocation of the
       corresponding setjmp macro
    3. that do not have volatile-qualified type
    4. and have been changed between the setjmp invocation and longjmp
       call
are indeterminate." (formatting and numbering added)

The C89 standard is similar (4.6.2.1) except that the second factor
above is not required in order for values to be indeterminate.

LLVM currently miscompiles the code
	int h = 0;
	...
	if (!l)
		siglongjmp(j, 1);
	h = 1;
	...
(note that h is an object of automatic storage duration, local to the
function, without volatile-qualified type, but is not modified until
*after* the longjmp call) into the following:
	int h = 0;
	...
	h = 1;
	if (!l)
		siglongjmp(j, 1);
	...
probably due to a lack of proper data dependency modelling in siglongjmp.

Our workaround replaces sigsetjmp/siglongjmp with a (not standard
compliant, but supported by clang) computed goto.

Reference:	https://llvm.org/bugs/show_bug.cgi?id=27190