Avoid concatenating multiple xattr read errors
This arose from running Tarsnap inside a jail:
    http://mail.tarsnap.com/tarsnap-users/msg01206.html

This patch doesn't address the root problem (FreeBSD not exposing xattr inside
a jail and/or Tarsnap not detecting it being in a jail and/or the user not
being able to specify a --no-xattr option (which will be introduced in
libarchive 3.2 but is not yet available in a stable release)).  But at least
the warning messages no longer look ridiculous.