Move SHA256 round constants into an array
This will allow for the "explicitly unrolled" sequence of rounds to
be replaced by a C loop in a future commit.

For diff minimization purposes, the RNDr macro still takes four
arguments; but the fourth argument (which used to be the relevant
round constant) is now ignored.

On {clang 3.8.0, gcc 4.8.5}/{-O2, -O3 -march=native}, the performance
impact of this is not measurably different from zero.