readpass.c: cast ECHO to tcflag_t
clang-3.8 treats ECHO as an int, which it performs a bitwise NOT upon, and then
implicitly casts to an unsigned int.  It's more clear to the compiler what's
happening if we treat ECHO as tcflag_t (an unsigned integer which is no greater
than the width of long) from the beginning.