Explicit cast fd to (size_t)
network_register(): we checked that fd >= 0 and fd < FD_SETSIZE [1].

network_deregister(): we checked that fd >= 0 and fd < callbacks_getsize(),
and we do not register more than FD_SETSIZE callbacks, so this cast is safe.

[1] Technically, FD_SETSIZE is the "Maximum number of file descriptors in an
fd_set structure."[2], which does not guarantee that this number will fit into
(size_t).  So I think that we could have a system with 64-bit (int), 32-bit
(size_t), for which FD_SETSIZE is not a safe check to ensure that we can cast
to (size_t).  However, since FD_SETSIZE is 1024 on my modern 64-bit OS, I
doubt any implementation will exceed 2,147,483,647.
[2] http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_select.h.html