Sanity-check for MGF1 mask length
Although the MGF1 hash function does not state an explicit absolute limit on
the size of its mask length, one can be inferred.  Part of the MGF1 hash
function is:

    3. For counter from 0 to \ceil (maskLen / hLen) - 1, do the following:
       a. Convert counter to an octet string C of length 4 octets (see
          Section 4.1):
             C = I2OSP (counter, 4) .
    https://tools.ietf.org/html/rfc3447#appendix-B.2.1

The I2OSP algorithm begins
    1. If x >= 256^xLen, output "integer too large" and stop.
    https://tools.ietf.org/html/rfc3447#section-4.1

Since xLen is 4, this means that our counter must be < 256^4 = 2^32.  In the C
code, our counter (pos) can reach a maximum value of (buflen - 1) / 32, and
UINT32_MAX is 2^32 - 1.