Explicit cast to (uint32_t)
In the tarsnap client, crypto_MGF1() is only ever called with buflen=32, 48, or
223.  So in that respect, the cast is safe.

However, the function claims to implement "The MGF1 mask generation function,
as specified in RFC 3447", and I can't see an absolute limit on the size in the
RFC.  There is a relative limit of:
    1. If maskLen > 2^32 hLen, output "mask too long" and stop.
    https://tools.ietf.org/html/rfc3447#appendix-B.2.1
but our hash function outputs 32 octets, so checking that maskLen (= buflen, in
the C code) is less than 2^32 * 32 is not going to help with ensuring that
(buflen / 32) will fit into (uint32_t).

Abandoning a theoretical limit and going for practical safety, we can work
backwards: In order for the existing code

    crypto_MGF1(..., size_t buflen) {
        size_t pos;
        uint32_t i;
        for (pos = 0; pos < buflen; pos += 32) {
            i = pos / 32;

to be safe, we should check that
    buflen / 32 <= UINT32_MAX
which is what I've done.