Explicit casts to (uint32_t) and sanity checks
`buflen` is passed as an argument from:
    tar/multitape/multitape_write.c	store_chunk()
which can be called from two places:
- callback_file(): this is called by chunkify_init(), which promises to create
  chunks no longer than the `maxlen` parameter.  We initialize this with
  MAXCHUNK (:= 261120 = 2^18-1024), so this will fit into (uint32_t).
- handle_chunk(): thin wrapper called by callback_h(), callback_t(), and
  callback_c().  These are all called from streams, which are initialized
  with stream_init(), which also uses chunkify_init() with MAXCHUNK as the
  `maxlen` parameter.
Therefore, provided that the chunk layer works, the `buflen` parameter will fit
into (uint32_t).  An extra sanity check was added for this.

`zlen` is the size of the compressed chunk.  The maximum compressed size is
defined in chunks_write_start(); after switching variables for constants:
    zbuflen = 1.001 * MAXCHUNK + 13 = 261394
If that is not large enough to hold the compressed data, zlib's compress2()
function will return Z_BUF_ERROR, which is caught by a switch() command.
Therefore `zlen` will fit into (uint32_t).

CHDATA_MALLOC and CHDATA_CTAPE are defined in
    tar/chunks/chunks_internal.h
and are (1 << 30) and (1 << 31), so these fit into (uint32_t), and will not
interfere with zlen (whose maximum value is less than 2^19).  An extra sanity
check was added to ensure that `zlen` is less than CHDATA_MALLOC (1 << 30) so
that it will not interfere with the extra flags.