Explicit cast to (size_t)
The handle for a sleeper is a (size_t); it is only treated as an (int) in the
API so that int network_sleep() can return -1.  As such, it is not valid to
pass a negative handle to network_desleep().

This adds a sanity check against a negative handle, which means that the later
cast to (size_t) is safe.