Explicit cast to (uint64_t) and (off_t)
multitape_write.c: Current filesystems have a limit of 2^64 bytes (or 16 EiB)
for a single file, but I added a sanity-check anyway for safety with future
operating systems.  You know, in case somebody wants to pay $2.3 billion /
month to Tarsnap in storage costs.  :)

multitape_read.c: Since we're safely storing an (off_t) in a (uint64_t), it's
safe to do the reverse.