Explicit cast to (ssize_t) in readtape_read*()
This file already did a
    CTASSERT(MAXCHUNK <= SSIZE_MAX);
so in readtape_readchunk(),
    len = le32dec(d->c.ch.len);
should already give us a value which is less than MAXCHUNK (and thus also less
than SSIZE_MAX).

A similar argument applies to readtape_read(); if we're reading chunk data, it
should be less than SSIZE_MAX.  If we're reading header data, there will be a
trivially small amount of data.  If we're reading trailer data, then it should
be at most MINCHUNK [= 4096] bytes.

However, I figured it wouldn't hurt anything to add sanity checks as well.

* Note that these casts are only safe due to the CTASSERT.  POSIX only
guarantees that SSIZE_MAX will be at least 32767, which is much smaller than
MAXCHUNK [= 261120].  Without the CTASSERT, a masochistic user (like "Mallory",
but directed at self-harm) could create archives on one machine, then attempt
to restore the archives on a POSIX system with a small SSIZE_MAX.