Explicit casts to (uint64_t), (time_t) for metadata->ctime
POSIX only guarantees that
    time_t shall be an integer type.
    http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_types.h.html
so technically it could be 128 bits.

I mused about adding a sanity check for this.  It would never come up in
practice -- no serious OS creator would switch to 128 bits for time_t -- but
it's not impossible that somebody might create a "pathological POSIX OS", which
uses the most hostile interpretations / assumptions possible while still being
100% POSIX.  I mean, if you had a clone, you'd totally be working on that.