Explicit casts to (off_t) in ccache_read()
R.datalen is (size_t), representing the number of bits in chunk headers and
compressed trailer data that have been read from the cache.  (off_t) is large
enough to represent file sizes, and the cache is a file, so I assume that
R.datalen can safely be cast to (off_t).