Set key to NULL when discarding old values.
In most cases we were about to immediately set the key to a new
value; but in crypto_keys_subr_import_HMAC we check the buffer
length before the malloc call.  As a result, reading a (corrupt)
tarsnap key file which has a value HMAC key followed by an HMAC
key with the same type but an invalid length resulted in the key
being freed in crypto_keys_subr_import_HMAC and then freed again
when keys were cleaned up during atexit handling.

Since the problem is triggered only by the handling of corrupt key
files and arises as a double free (or access to freed memory, but
the memory in question won't have been overwritten) moments before
we exit(1), this is pretty much harmless; but it is *technically*
a crash, and could have been worse if the code had been structured
differently.

Reported by:	GwanYeong Kim
Security:	Access to freed memory / double-free during error
		exit path.
Bug Bounty:	$100