Add --dump-config option
Occasionally people get confused because they forgot that they'd configured
tarsnap to behave in a certain way, and/or they didn't realize there was a
default config file (particularly if tarsnap was installed from a package).

This prints out every config option that the tarsnap binary receives, and where
it came from (i.e. config file, command line).  We can then add it to the FAQ:
    Q: Tarsnap is behaving weirdly!
    A: Try running the command with the --dump-config option to check that it's
       receiving the correct options.  Sometimes people forget they configured
       tarsnap a certain way months or years ago!

I considered (and mostly-implemented) printing the command-line options during
the actual parsing, so that it would be split into
    --print-stats
    -f blah
    --csv-file blah.log
but that code ended up being rather ugly.  As far as "helping users debug their
setup" goes, there's no need to be clever about this.

I definitely think that we need to "echo" the command-line, though, as that
will clear up any shell expansion.  For example:
    $ ./tarsnap --dump-config --print-stats -f *
    Command-line:
      ./tarsnap --dump-config --print-stats -f caps check-iwyu.sh check.sh config.h config.log config.status cpusupport-config.h err-1000 err-100000 iwyu.log keygen keymgmt keyregen lib libarchive libcperciva Makefile recrypt stamp-h1 tar tarsnap tarsnap-keygen tarsnap-keymgmt tarsnap-keyregen tarsnap-recrypt valgrind-full.log valgrind.supp
    Reading from config file: /home/td/.tarsnaprc
    ...
clearly demonstrates the user's problem.