scryptdec_file(): refactor with scryptdec_file_prep()
This has no effect on external code.

This commit includes scrypt's "fix memleak if the input file can't be processed"
commit that occurred when:
- the input filename successfully responded to open()
- the input file could not be read, or contained the wrong magic header,
  or did not contain the full 96-byte header.

It is safe to call scryptdec_file_cookie_free() on C, even if C->dk hasn't been
initialized, because C->dk is not dynamically allocated.  (We could technically
use free(C) instead of scryptdec_file_cookie_free(), but I thought this fix was
clearer to read.)