crypto_compat: free OpenSSL error queue
This frees memory that valgrind reports as being "still reachable" when tarsnap
creates a new keyfile (i.e. if running
    tarsnap --no-default-config --dry-run -c .
).  That memory arises from our use of RSA_generate_key_ex().

ERR_remove_state() in 1.0.2:

    Since error queue data structures are allocated automatically for new
    threads, they must be freed when threads are terminated in order to avoid
    memory leaks.
    https://www.openssl.org/docs/man1.0.2/crypto/ERR_remove_state.html

ERR_remove_thread_state() in 1.1.0:

    HISTORY

    ERR_remove_state() was deprecated in OpenSSL 1.0.0 when
    ERR_remove_thread_state() was introduced.

    ERR_remove_thread_state() was deprecated in OpenSSL 1.1.0 when the thread
    handling functionality was entirely rewritten.
    https://www.openssl.org/docs/man1.1.0/crypto/ERR_remove_thread_state.html

I wish that we didn't have to deal with internal OpenSSL implementation issues,
but I guess that's just the state of the industry.