Zero HMAC-SHA256 context in _Final.
This zeroing was lost in commit 40502c96 (May 24, 2016) when most
of the functions in this file split into versions which performed
zeroing and internal function which did not.  Unfortunately, the
HMAC_SHA256_Final function had relied upon SHA256_Final to zero
its context, but was changed the use the internal (non-zeroing)
function and the necessary zeroing was not added.

Reported by:	Graham Percival