memlimit: don't use RLIMIT_DATA if we HAVE_MMAP
The memlimit code serves two purposes:
1. Making sure that we don't try to use more memory than we can.
2. Trying to avoid using more memory than we should.

On systems without mmap, scrypt uses malloc and RLIMIT_DATA is likely to
be a hard limit on what we can allocate; so in that case we respect the
value.

On systems with mmap, RLIMIT_DATA has historically still been a useful
indication of how much memory "well behaved" programs should use; but as
recent mallocs have switched from using sbrk to using mmap(MAP_ANON) it
has become common for malloc to not respect RLIMIT_DATA.  As a result,
there are now systems where RLIMIT_DATA has "fallen behind" -- with
scrypt being the only casualty, since nobody else ever checks the value
explicitly.

By looking at RLIMIT_DATA only in the !HAVE_MMAP case, we respect it
when it is important while ignoring it in the cases where it is liable
to be meaningless.