siginfo: swap printing overall info and specific files
Swap the order of the new output message added in 3ee1bfbd4f7cd4dd09bd3f2767207f194b16824a.

We currently output SIGINFO / SIGUSR1 messages in this form:

    adding random-300k (0 / 300000 bytes)
    Processed 1 files, 300544 bytes

where both lines are reporting about the process at the same time-point.
However, the two lines are not actually temporarily equal: the second
line reports on the past, while the first line reports on the present.
That can get confusing when there's multiple messages.  For example,
consider what happens if we add `raise(SIGUSR1);` to trigger after
recording that we're processing each file (i.e. line 1002 of write.c).

    $ ./tarsnap --dry-run -c 1 2
    adding 1 (0 / 1234 bytes)
    adding 2 (0 / 1234 bytes)
    Processed 1 files, 2048 bytes

In this case, the flow of time is:
    t_0 present
    t_1 present
    t_1 past

If we swap the order of "Processed" and "adding", then it makes much
more sense:

    $ ./tarsnap --dry-run -c 1 2
    adding 1 (0 / 1234 bytes)
    Processed 1 files, 2048 bytes
    adding 2 (0 / 1234 bytes)

The flow of time is now:
    t_0 past
    t_1 past
    t_1 present