bsdtar: bail on --dry-run given a bad passphrase
Previously,
- if we're doing a --dry-run,
- given a passphrase-protected keyfile,
- specified in a config file (but not the command-line),
- the user fails to provide the correct passphrase,
tarsnap printed the
    Ignoring cachedir due to missing or invalid keyfile in config file.
message and continued processing the --dry-run.

This commit changes that behaviour to bail instead.  Note that "bail for
a bad passphrase" is already the behaviour if the --keyfile is specified
on the command-line; the change is that now tarsnap behaves the same way
with a bad passphrase, regardless of whether the keyfile was specified
on the command-line or config file.

This arose on 2016-01-06, when we added
    Allow dryrun with an unusable keyfile in config
    dd4558f088f19d595fb2aa801124d18d6c76e787
so that people could generate --dry-run stats without having to modify
the default /usr/local/etc/tarsnap.conf (and without adding
--no-default-config).