tar/: add "platform.h" to all .c files
This fixes the client being broken on 32-bit platforms (i386 and ARM).

The problem is that I was unaware that glibc processed the "feature test
macros" [1] the first time the compiler encountered any #include
statements for system libraries [2].  As a result, when I removed the
"unnecessary" bsdtar_platform.h includes [3], that meant that some
object files interpreted off_t and ino_t as being 4 bytes, while other
object files interpreted them as being 8 bytes.  That led to all sorts
of Bad Things (tm) happening.

Fortunately, one of those Bad Things (tm) was that the networking code
could not read the machine number correctly, so the client could not
communicate with the server.  No user data was at risk from anybody
using the unreleased git master.

If we add more files to tar/, or add any library code which uses off_t
or ino_t, we'll need to add "platform.h" to more files.  It's easy to
check this in git [4].

[1] https://www.gnu.org/software/libc/manual/html_node/Feature-Test-Macros.html
[2] I tested extensively on amd64, as did Travis-CI, but I neglected to
    test on a 32-bit system.
[3] remove-unnecessary-bsdtar-h branch: 1f56dc7c35b32d9ce37d727215c1293e59d0f2e1

[4] We can check which .c files contain platform.h with:
        $ git grep "platform.h" -- "tar/*.c"
    This also detects the files which include bsdtar_platform.h.  Note
    that git interprets "tar/*.c" as recursing into all subdirectories.

    Conversely, we can check which .c files do *not* contain platform.h with:
        $ git grep -L "platform.h" -- "tar/*.c"
    At the moment, this only includes tar/getdate.c, which is an
    autogenerated file which we do not want to modify.