Allow --dump-config to be a mode (like --print-stats)
We occasionally have bug reports which boil down to "I didn't realize
that tarsnap was looking at a global config file".  If we suspect such a
scenario, it would be nice if we could ask people to run
    tarsnap --dump-config
and see something like:

    Command-line:
      ./tarsnap --dump-config
    Reading from config file: /home/td/.tarsnaprc
      nodump
      ...
    Reading from config file: /usr/local/etc/tarsnap.conf
      cachedir /usr/local/tarsnap-cache
      keyfile /root/tarsnap.key
      ...

Previously, we'd have to ask them to send us the results of
    tarsnap --dump-config --print-stats
"but for privacy reasons, don't include the statistics".  We could
replace --print-stats with --list-archives or -c --dry-run /dev/null or
something like that... but if we're dealing with a confused user, then
having a simple diagnostic would be nice.