POSIX: import posix-l.c and posix-l.sh
When we imported posix-cflags.sh in 9fe9ace5ee8da640ecb154a6f624a010a667ec6a,
we omitted posix-l.sh because "autotools seems to do a decent job of
figuring it out".

That reason is still valid, but in the interest of maintability I now
think that it makes sense to include both posix-cflags.sh and posix-l.sh
together, even if we choose not to use LDADD_POSIX in Makefile.am.