man: alphabetical order for --newer-than, --newer-mtime-than
These were added in tarsnap 1.0.0 41f175dbc884f04e36a98811188e1100bb4dbfae
with --newer-than first, probably because --newer-mtime-than is best
understood as a modification of --newer-than.

However, this pair are the only exception to the ordering [1] of the
OPTIONS section, and they're right next to each other anyway.  I'd
rather swap them so that there's no exceptions.

[1] alphabetical order, ignoring case unless it's a tie-breaker.