getopt: rename "maximum number of options" variable
Naming this argument `ln` was slightly misleading, because it's actually
the difference between line numbers (namely, `ln - getopt_ln_min` from
GETOPT_DEFAULT), not a line number (which is what `ln` means elsewhere
in this file and the header).

But more to the point, in getopt_setrange(), that value represents "the
maximum number of options", so the name should give a clue about that.