tests: check -c, -d with a real keyfile
The key provided by $TARSNAP_TEST_KEYFILE should only be used for the
automatic test, and should not have any archives stored "at rest".
(This particular test does not involve --nuke, but later one(s) might.)

This test will:
- skip if there's no $TARSNAP_TEST_KEYFILE
- error if there's anything in --list-archives
- initialize a cachedir
- fsck
- create an archive
- delete an archive
- check that we have no data left on the server

Most of these tests use the typical "PASS / FAIL" output.  However, if
there's an error with --list-archives or anything in that list, we exit
with a more specific error message, as this may indicate an incorrect
keyfile being passed to $TARSNAP_TEST_KEYFILE.