tests: avoid special parameter
We could have replaced $@ with "$*" -- provided that IFS was not set to an
awkward value [1].  In this function, all we want is a string, so let's
just avoid that whole mess.

This commit may require a slight change to project code: the old format
of finding a binary which supports arguments:
    find_system scrypt enc -P
must be changed to:
    find_system scrypt "enc -P"

This is slightly less idiomatic, but results in much simpler code.

[1] "$*" works if IFS is unset, or the first character in IFS is a space
    or a tab.  I would be stunned if we found a shell where that wasn't
    true, but why risk it?
    https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_05_03

Reported by:	shellcheck