tests/valgrind: don't include ".log" in valgrind log basename
This is preparation for handling forked programs in our valgrind tests.
Currently, our valgrind logfiles are in the form:
    01-crc32-00.log
but we're moving towards including the pid:
    01-crc32-00-7114.log

Stripping the ".log" will allow us to use a glob, so that we don't need
to know the pid to get the relevant filename.