Don't close stdin when cleaning up after a --dry-run
I broke this in:
    2015-07-30 Allow --dry-run compress without a keyfile
    92d5c36d72de999836d124f9983c8643e6123aa3

If we're doing a --dry-run, then writetape_open()'s

    /* If we're using a cache, lock the cache directory. */
    if ((cachedir != NULL) && ((d->lockfd = multitape_lock(cachedir)) == -1))
        goto err3;

does not set d->lockfd.  That variable was initialized with

    memset(d, 0, sizeof(struct multitape_write_internal));

so as a result, a --dry-run was mysteriously closing stdin.

These checks are not necessary for multitape_delete.c and multitape_fsck.c,
because we can't avoid using the cache for those modes.