crypto_aesctr_aesni: add
To keep the history of this branch as clear as possible,
crypto_aesctr_aesni_stream() is an exact copy of crypto_aesctr_stream().

The "helper" functions in crypto_aesctr_shared.c are defined as
"inline", and for performance reasons that is crucial -- if the compiler
doesn't inline them, we face an immediate 33% speed penalty on x86.  For
that reason, at the top of crypto_aesctr_stream(), we either immediately
jump to crypto_aesctr_aesni_stream(), or stay completely with the
generic code.

In the future, crypto_aesctr_aesni_stream() will be optimized, but only
when generating & using a whole cipherblock (16 bytes) at once.  As
such, there's no point switching to that function unless buflen >= 16.