Replace --passphrase-stdin with --passphrase method:arg
This commit removes the --passphrase-stdin option.  That was added in
    2018-12-28 Add --passphrase-stdin
    d577115b29a0aac7419b4c2c42eacf4b56aebd00
which was after 1.0.39 (2017-07-28).  There have been no official
releases of tarsnap which included --passphrase-stdin.

Implementation note: we need to add `str` to dooption() and do a
strdup(str) because passphrase_entry_parse() makes a (const char *)
point to the passphrase_arg portion of conf_arg, but this is only a
temporary variable.  The dooption() function already contains multiple
other strdup() calls, for precisely this reason.

An alternate method would be to change passphrase_entry_parse() to
allocate a new string for the (const char **), but that would unduly
complicate that function (and by extension, complicate scrypt).