crypto_aes_arm.c: add SubWord_duplicate()
In the odd-numbered rounds of generating a 256-bit AES key, we only need
to use SubWord.  We can do this with the AESE instruction, provided that
we can negate the AddRoundKey and ShiftRows operations.
- negating AddRoundKey is done by providing a key of zeros
- negating ShiftRows is done by making four copies of the uint32_t that
  we wish to perform SubWord upon.  I learned about this trick in [1].

[1] https://community.arm.com/developer/tools-software/tools/b/tools-software-ides-blog/posts/porting-putty-to-windows-on-arm

There is a temporary extra instruction here: in the odd rounds,
SubWord_duplicate() takes care of duplicating the 32-bit word to all
lanes, but the next line of MKRKEY256 is
    _t = vdupq_laneq_u32_u8(_t, lane);
This duplication is still necessary for the even rounds, and it's
harmless for the odd rounds.  It will be removed in the following
commit, once we've added a _duplicate() function call for the even
rounds.