crypto_aes_arm.c: add SubWord_RotWord_XOR_duplicate()
This moves "duplicate the desired 32-bit word to all lanes in the
128-bit vector" into the _mm_aeskeygenassist_si128() function.  As such,
it is no longer an accurate shim for the x86 AESNI instruction of that
name.  I've therefore renamed it to SubWord_RotWord_XOR_duplicate().

This commit also removes the vdupq_laneq_u32_u8() calls in MKRKEY128 and
MKRKEY256, as this is handled by SubWord_duplicate() and
SubWord_RotWord_XOR_duplicate().  The previous commit message noted as
being unnecessary for odd rounds; it is now unnecessary for even rounds
as well.