memlimit.c: remove implicit maxmemfrac limit, and allow up to 1.0
Previously, memtouse() would silently treated any ${maxmemfrac} value
greater than 0.5 as 0.5, and would also treat a 0 as 0.5.

This commit:
- removes that implicit handling; maxmemfrac must be specified explicitly.
- sets the upper limit to 1.0.

Assuming that the memory limit detection is accurate [1], I don't know
why anybody would want to set maxmemfrac to 1.0.  But that's the
business of the client code, not this library function.  (We have no
plans to allow the scrypt binary to set this to more than 0.5.)

[1] There's no reliable cross-platform way to detect the memory limits,
    so it's entirely possible that it won't work on a new release of an
    operating system.  For example, see
        2020-05-18 memlimit: USE_RLIMIT_DATA on OpenBSD
        b77dfb2bb26f94c34ade9bb5d08b985046e7cdbb