build: allow users to manually specify POSIX_SH
All POSIX-compatible systems are required to provide:
    command -p getconf PATH
    https://pubs.opengroup.org/onlinepubs/9699919799/utilities/command.html
    https://pubs.opengroup.org/onlinepubs/9699919799/utilities/getconf.html

However, the whole point of the configure script is to support buggy
systems.  As such, we might as well allow users to manually specify the
location of a POSIX-compliant shell, in case there's any system which
does not include `command` and/or `getconf`.