chunks_directory.c: better check for the file size
This should have been part of
    2008-04-25 Tarsnap 1.0.0
    41f175dbc884f04e36a98811188e1100bb4dbfae

Due to this missing size check, we used
    sb.st_size - sizeof(struct chunkstats_external))
without checking for underflow.  In particular,
sizeof(struct chunkstats_external) is 24, so we had potential problems
if sb.st_size was 0, 1, 2... 23.

Most of these values would be caught by the modulo check; the only
exception is if the original file size was 8 bytes, in which case we
would compute
    numchunks = (size_t)(-16) / 48 == SIZE_MAX / 48

The next check (for an integer overflow) will flag that value as being a
problem, but only because
    sizeof(struct chunkdata_statstape) >= sizeof(struct chunkdata_external)
Also, we would exit with the misleading message:
    warn0("on-disk directory is too large: %s", s);

In summary: if the chunks directory file was too small, we would still
exit with an error.  But it wasn't the right error message, and it does
not appear to have been deliberate.

Reported by:	Rasmus Villemoes
Bug bounty:	$50