mkdebsource.sh: adjust for X.Y.Z.A release numbers
We previously used a lettered suffix (e.g., "a", "b") to indicate
pre-release version numbers.

Reminder: we don't need to run this script if we have a new tarsnap
version number.  It's only needed if we're making a second (or later)
debian package for the same tarsnap version, i.e. if there's a problem
in 1.0.39.99-1 which requires a 1.0.39.99-2 which only has changes to
the pkg/debian/ directory and absolutely nothing else.