tests: delete unnecessary *.{exit,desc} files
Previously, we would have dozens of "*.exit" and "*.desc" files in the
tests-output/ directory.  If the checks completed successfully, those
files contributed nothing of value, so we will now delete them.

This commit does not delete those files for any check which failed.
As a result, it will be much easier to find & investigate the specific
failure(s).

We still keep all the files which are created by specific tests (e.g.,
01-crc32-stdout.txt, 02-getopt-stdout.txt).