Actions: use openssl 3.0 on macOS, plus minor tweaks
Github actions now adds

    CPPFLAGS=-I/usr/local/opt/openssl/include

to the environment, which points to openssl 3.0 rather than openssl@1.1
which we're explicitly asking for.

We _could_ get into weird fights about the order of -I statements
(controlled via environment variables such as CPPFLAGS_OSX) in order to
force clang to find the appropriate headers, but that feels like would
be getting increasingly fragile and not worth it.