write.c: include padding in final progress message
When I added the "final progress notification" siginfo_printinfo() in:
    2019-10-15 Allow a final progress notification
    763fa25c8a3c8beeeb0a28126f8dd5115c31e9c7
I was mistakenly concerned that archive_write_close() might cause
archive_file_count(a) or archive_position_uncompressed(a) to report
inaccurate information, so I put the notification above that.

Unfortunately, that meant that the final progress message, which is
always printed if we were given --progress-bytes X:
    2019-10-15 Add a final "Processed" message for --progress-bytes
    7d6171aa501ab7a9644b57f0205255c877a114a2
didn't include the final 2 padding blocks of 512 bytes each.